% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bcbio_nfcore_check}
\alias{bcbio_nfcore_check}
\title{Function to check samplesheet for nf-core}
\usage{
bcbio_nfcore_check(file)
}
\arguments{
\item{file}{path to CSV file for nf-core}
}
\description{
Function to check samplesheet for nf-core
}
\examples{

bcbio_nfcore_check(system.file("extdata", "rnaseq_good.csv", package = "bcbiR") )

}
