% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bcbio_set_project}
\alias{bcbio_set_project}
\title{Function to help with project name used for parent folder}
\usage{
bcbio_set_project()
}
\value{
A string list with hbc_code, and project folder name
}
\description{
This function will ask for user input about:
\itemize{
\item numeric code
\item PI full name
\item technology
\item tissue
\item organism
\item project description
}
}
\details{
It removes special character with \verb{_}. The output is a guideline to
what the folder used can be.
}
