% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcbio_vsd_data}
\alias{bcbio_vsd_data}
\title{varianceStabilizingTransformation Object from Test Data}
\format{
\subsection{\code{bcbio_vsd_data}}{

VSD object from DESeq2 data
}
}
\usage{
bcbio_vsd_data
}
\description{
An example of the output of varianceStabilizingTransformation to
show different visualization code
}
\keyword{datasets}
