% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\docType{data}
\name{cb_friendly_colors}
\alias{cb_friendly_colors}
\title{list of colorblind-friendly colors, developed by James and client}
\format{
An object of class \code{character} of length 16.
}
\usage{
cb_friendly_colors
}
\description{
list of colorblind-friendly colors, developed by James and client
}
\keyword{datasets}
