% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{cb_friendly_pal}
\alias{cb_friendly_pal}
\title{access cb friendly palette by name, reversing if necessary}
\usage{
cb_friendly_pal(palette = "main", reverse = F, ...)
}
\arguments{
\item{palette}{name of the palette to be returned}

\item{reverse}{boolean, reverse order of colors in palette}
}
\description{
access cb friendly palette by name, reversing if necessary
}
