% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\docType{data}
\name{cb_friendly_palettes}
\alias{cb_friendly_palettes}
\title{define main colorblind-friendly palette as well as sub-palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
cb_friendly_palettes
}
\description{
define main colorblind-friendly palette as well as sub-palettes
}
\keyword{datasets}
