% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_friendly.R
\name{scale_fill_cb_friendly}
\alias{scale_fill_cb_friendly}
\title{use cb friendly colors as fill aesthetic with ggplot}
\usage{
scale_fill_cb_friendly(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of the palette to be returned (main, cool, hot, grey, white_to_blue, or heatmap)}

\item{discrete}{boolean, whether to make palette discretely divided into colors or continuous}

\item{reverse}{boolean, reverse order of colors in palette}

\item{...}{pass to ggplot}
}
\description{
use cb friendly colors as fill aesthetic with ggplot
}
