% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_library}
\alias{use_library}
\title{Set custom library to your project}
\usage{
use_library(custom_lib_path, project_path = ".")
}
\arguments{
\item{custom_lib_path}{path with a R library folder}

\item{project_path}{project path}
}
\description{
This function will create the .Rprofile file in a project path
to use a custom library path for future session
}
\examples{
path <- withr::local_tempdir()
use_library(.libPaths(), path)
}
