% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbioR-package.R
\docType{package}
\name{bcbioR-package}
\alias{bcbioR}
\alias{bcbioR-package}
\title{bcbioR: Templates and functions to guide downstream analysis and data interpretation}
\description{
Collaborative code repository at the Harvard Chan Bioinformatics Core.
}
\author{
\strong{Maintainer}: Pantano Lorena \email{lorena.pantano@gmail.com}

}
\keyword{internal}
